refactor: 迁移到 Reown AppKit，Pair 优化 reserve 打包存储
- web: RainbowKit 换成 Reown AppKit，清理残留引用和死代码
- contracts: SimplePair reserve0/1 打包进 uint112+uint32，加 TWAP 累积字段
- 补充注释，统一 revert 信息

+++ contracts/src/SimpleRouter.sol
@@ -9,7 +9,8 @@ import {SimpleLibrary} from "./libraries/SimpleLibrary.sol";
 
 /**
  * @title SimpleRouter
- * @notice 用户交互的主入口：添加/移除流动性、swap
+ * @notice 用户交互的主入口：添加/移除流动性、swap。自身不持有任何资产，
+ *         只负责「计算 + 转账 + 调用 Pair」的封装。
  */
 contract SimpleRouter {
     using SafeERC20 for IERC20;
@@ -40,7 +41,7 @@ contract SimpleRouter {
             return (0, 0);
         }
         (address token0,) = SimpleLibrary.sortTokens(tokenA, tokenB);
-        (uint256 reserve0, uint256 reserve1) = SimplePair(pair).getReserves();
+        (uint112 reserve0, uint112 reserve1,) = SimplePair(pair).getReserves();
         (reserveA, reserveB) = tokenA == token0 ? (reserve0, reserve1) : (reserve1, reserve0);
     }
 
@@ -48,7 +49,7 @@ contract SimpleRouter {
         return _getReserves(tokenA, tokenB, false);
     }
 
-    // 计算多跳 swap 的输出
+    /// @notice 沿路径计算多跳 swap 的每一跳输出（简化版通常只用 path.length == 2）
     function getAmountsOut(uint256 amountIn, address[] calldata path) public view returns (uint256[] memory amounts) {
         require(path.length >= 2, "Invalid path");
         amounts = new uint256[](path.length);
@@ -59,7 +60,7 @@ contract SimpleRouter {
         }
     }
 
-    // 计算多跳 swap 的输入
+    /// @notice 沿路径反推多跳 swap 需要的每一跳输入
     function getAmountsIn(uint256 amountOut, address[] calldata path) public view returns (uint256[] memory amounts) {
         require(path.length >= 2, "Invalid path");
         amounts = new uint256[](path.length);
@@ -70,7 +71,7 @@ contract SimpleRouter {
         }
     }
 
-    // 添加流动性
+    /// @notice 添加流动性：pair 不存在则先创建，再按当前比例（或用户期望值）转入两种代币铸 LP
     function addLiquidity(
         address tokenA,
         address tokenB,
@@ -106,7 +107,7 @@ contract SimpleRouter {
         liquidity = SimplePair(pair).mint(to);
     }
 
-    // 移除流动性
+    /// @notice 移除流动性：把 LP token 转入 pair 自身销毁，换回按比例的两种代币
     function removeLiquidity(
         address tokenA,
         address tokenB,
@@ -127,7 +128,7 @@ contract SimpleRouter {
         require(amountA >= amountAMin && amountB >= amountBMin, "Insufficient output");
     }
 
-    // Swap：精确输入
+    /// @notice Swap：精确输入换尽量多的输出
     function swapExactTokensForTokens(
         uint256 amountIn,
         uint256 amountOutMin,
@@ -141,7 +142,7 @@ contract SimpleRouter {
         _swap(amounts, path, to);
     }
 
-    // Swap：精确输出
+    /// @notice Swap：精确输出，投入不超过 amountInMax
     function swapTokensForExactTokens(
         uint256 amountOut,
         uint256 amountInMax,
@@ -160,7 +161,8 @@ contract SimpleRouter {
             (address input, address output) = (path[i], path[i + 1]);
             (address token0,) = SimpleLibrary.sortTokens(input, output);
             uint256 amountOut = amounts[i + 1];
-            (uint256 amount0Out, uint256 amount1Out) = input == token0 ? (uint256(0), amountOut) : (amountOut, uint256(0));
+            (uint256 amount0Out, uint256 amount1Out) =
+                input == token0 ? (uint256(0), amountOut) : (amountOut, uint256(0));
             address to = i < path.length - 2 ? _pairFor(output, path[i + 2]) : finalTo;
             SimplePair(_pairFor(input, output)).swap(amount0Out, amount1Out, to);
         }
